fixing a few VS warnings
Former-commit-id: 3932ad20d230e3e952521ca2d3fd5804f4caaa23

+++ modules/SofaBaseCollision/SphereModel.inl
@@ -112,7 +112,7 @@ void TSphereModel<DataTypes>::draw(const core::visual::VisualParams* vparams,int
 {
     TSphere<DataTypes> t(this,index);
 
-	vparams->drawTool()->drawSphere(t.p(), t.r());
+    vparams->drawTool()->drawSphere(t.p(), (float)t.r());
 }
